// libs/core/dtos/pagination.dto.ts
/** Pagination meta info. */
export type PaginationDto<T> = {

	/** Total count of items. */
	readonly count: number;

	/** Next page of items. */
	readonly next: string;

	/** Previous page of items. */
	readonly previous: string;

	/** Array of items requested. */
	readonly results: readonly T[];
};

/** Pagination list cursor. */
export type PaginationListCursorDto<T> = {

	/** Next page of items. */
	readonly next: string | null;

	/** Previous page of items. */
	readonly previous: string | null;

	/** Array of items requested. */
	readonly results: readonly T[];
};